Shift rime.py around slightly in order to have the main method be callable externally (when rime.py is imported into rime_main.py)
Added connections from rime_main.py to the backend, so both the RIME CLI and
GUI can all be run from the same python file.

Redirected standard out in the GUI to the Input statistics text box - this
is only for proof of concept.  Next step is attaching the rest of the methods
in rime.py to the GUI.